// src/schemas.ts
import { z } from "zod"

export const MealSchema = z.object({
  idMeal: z.string(),
  strMeal: z.string(),
  strMealAlternate: z.string().nullable(),
  strCategory: z.string(),
  strArea: z.string(),
  strInstructions: z.string(),
  strMealThumb: z.string(),
  strTags: null,
  strYoutube: z.string().nullable(),
  strIngredient1: z.string(),
  strIngredient2: z.string(),
  strIngredient3: z.string(),
  strIngredient4: z.string(),
  strIngredient5: z.string(),
  strIngredient6: z.string(),
  strIngredient7: z.string(),
  strIngredient8: z.string(),
  strIngredient9: z.string(),
  strIngredient10: z.string(),
  strIngredient11: z.string(),
  strIngredient12: z.string(),
  strIngredient13: z.string(),
  strIngredient16: z.string(),
  strIngredient17: z.string(),
  strIngredient18: z.string(),
  strIngredient19: z.string(),
  strIngredient20: z.string(),
  strMeasure1: z.string(),
  strMeasure2: z.string(),
  strMeasure3: z.string(),
  strMeasure4: z.string(),
  strMeasure5: z.string(),
  strMeasure6: z.string(),
  strMeasure7: z.string(),
  strMeasure8: z.string(),
  strMeasure9: z.string(),
  strMeasure14: z.string(),
  strMeasure15: z.string(),
  strMeasure16: z.string(),
  strMeasure17: z.string(),
  strMeasure18: z.string(),
  strMeasure19: z.string(),
  strMeasure20: z.string(),
  strSource: z.string(),
  strImageSource: z.string().nullable(),
  strCreativeCommonsConfirmed: z.boolean().nullable(),
  dateModified: z.string(),
})

export type Meal = z.infer<typeof MealSchema>